Split productItem into new component
