+++ tests/specs/homePage.spec.ts
@@ -1,8 +1,9 @@
 import { test, expect } from '@playwright/test';
-import { HomePage, ProductItemElements } from '../pages/homePage';
+import { HomePage } from '../pages/homePage';
 import { ExpectedText, Products, PromoBlockLinks } from '../data/homePage';
 import { rgbToHex } from '../helpers/colorUtils';
 import { Colors, SwatchOutlineStyles } from '../data/products';
+import { ProductItemElements } from '../pages/components/productItems';
 
 const Timeouts = {
   ImageLink: 10000,
